+++ System/OsString.hs
@@ -14,7 +14,7 @@
 --
 -- It captures the notion of syscall specific encoding (or the lack thereof) to avoid roundtrip issues
 -- and memory fragmentation by using unpinned byte arrays. Bytes are not touched or interpreted.
-module System.OsString
+module System.OsString {-# DEPRECATED "Use System.OsString from os-string package instead. This module will be removed in filepath >= 1.5." #-}
   (
   -- * String types
     OsString
@@ -126,81 +126,5 @@ module System.OsString
   )
 where
 
-import System.OsString.Internal
-    ( unsafeFromChar
-    , toChar
-    , encodeUtf
-    , encodeWith
-    , encodeFS
-    , osstr
-    , pack
-    , empty
-    , singleton
-    , decodeUtf
-    , decodeWith
-    , decodeFS
-    , unpack
-    , snoc
-    , cons
-    , last
-    , tail
-    , uncons
-    , head
-    , init
-    , unsnoc
-    , null
-    , length
-    , map
-    , reverse
-    , intercalate
-    , foldl
-    , foldl'
-    , foldl1
-    , foldl1'
-    , foldr
-    , foldr'
-    , foldr1
-    , foldr1'
-    , all
-    , any
-    , concat
-    , replicate
-    , unfoldr
-    , unfoldrN
-    , take
-    , takeEnd
-    , takeWhileEnd
-    , takeWhile
-    , drop
-    , dropEnd
-    , dropWhileEnd
-    , dropWhile
-    , break
-    , breakEnd
-    , span
-    , spanEnd
-    , splitAt
-    , split
-    , splitWith
-    , stripSuffix
-    , stripPrefix
-    , isInfixOf
-    , isPrefixOf
-    , isSuffixOf
-    , breakSubstring
-    , elem
-    , find
-    , filter
-    , partition
-    , index
-    , indexMaybe
-    , (!?)
-    , elemIndex
-    , elemIndices
-    , count
-    , findIndex
-    , findIndices
-    )
-import System.OsString.Internal.Types
-    ( OsString, OsChar )
+import "os-string" System.OsString
 import Prelude ()