+++ System/OsString/Posix.hs
@@ -1,7 +1,4 @@
-{-# LANGUAGE CPP #-}
-#undef WINDOWS
-#define MODULE_NAME     Posix
-#define PLATFORM_STRING PosixString
-#define PLATFORM_WORD   PosixChar
-#define IS_WINDOWS      False
-#include "Common.hs"
+module System.OsString.Posix {-# DEPRECATED "Use System.OsString.Posix from os-string package instead. This module will be removed in filepath >= 1.5." #-} (module P) where
+
+import "os-string" System.OsString.Posix as P
+