+++ tests/bytestring-tests/Properties/OsString.hs
@@ -1,7 +1,0 @@
-{-# LANGUAGE CPP #-}
-#undef WORD16
-#undef POSIX
-#undef WIN
-#define OSWORD
-#include "Common.hs"
-