+++ tests/bytestring-tests/Properties/PosixString.hs
@@ -1,7 +1,0 @@
-{-# LANGUAGE CPP #-}
-#undef WORD16
-#define POSIX
-#undef WIN
-#undef OSWORD
-#include "Common.hs"
-