+++ tests/bytestring-tests/Properties/ShortByteString.hs
@@ -1,7 +1,0 @@
-{-# LANGUAGE CPP #-}
-#undef WORD16
-#undef WIN
-#undef POSIX
-#undef OSWORD
-#include "Common.hs"
-