+++ tests/bytestring-tests/Properties/ShortByteString/Word16.hs
@@ -1,6 +1,0 @@
-{-# LANGUAGE CPP #-}
-#define WORD16
-#undef WIN
-#undef POSIX
-#undef OSWORD
-#include "../Common.hs"