+++ tests/bytestring-tests/Properties/WindowsString.hs
@@ -1,7 +1,0 @@
-{-# LANGUAGE CPP #-}
-#define WORD16
-#define WIN
-#undef POSIX
-#undef OSWORD
-#include "Common.hs"
-