+++ tests/filepath-tests/TestGen.hs
@@ -2,6 +2,7 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE ViewPatterns #-}
 {-# LANGUAGE CPP #-}
+{-# LANGUAGE PackageImports #-}
 {-# OPTIONS_GHC -Wno-name-shadowing #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 module TestGen(tests) where
@@ -14,11 +15,11 @@ import Data.String
 import GHC.IO.Encoding.Failure ( CodingFailureMode(..) )
 import GHC.IO.Encoding.UTF16 ( mkUTF16le )
 import GHC.IO.Encoding.UTF8 ( mkUTF8 )
-import System.OsString.Internal.Types
-import System.OsPath.Encoding.Internal
+import "os-string" System.OsString.Internal.Types
+import "os-string" System.OsString.Encoding.Internal
 import qualified Data.Char as C
-import qualified System.OsPath.Data.ByteString.Short as SBS
-import qualified System.OsPath.Data.ByteString.Short.Word16 as SBS16
+import qualified "os-string" System.OsString.Data.ByteString.Short as SBS
+import qualified "os-string" System.OsString.Data.ByteString.Short.Word16 as SBS16
 import qualified System.FilePath.Windows as W
 import qualified System.FilePath.Posix as P
 import qualified System.OsPath.Windows as AFP_W